+++ src/main/java/in/koreatech/koin/domain/timetable/repository/TimetableFrameRepository.java
@@ -9,21 +9,22 @@ import in.koreatech.koin.domain.timetable.exception.TimeTableNotFoundException;
 import in.koreatech.koin.domain.timetable.model.TimetableFrame;
 
 public interface TimetableFrameRepository extends Repository<TimetableFrame, Integer> {
+
     List<TimetableFrame> findAllByUserId(Integer userId);
 
     List<TimetableFrame> findAllByUserIdAndIsMain(Integer userId, Boolean isMain);
 
     Optional<TimetableFrame> findById(Integer id);
 
-    Optional<TimetableFrame> findByUserIdAndSemester(Integer userId, String semester);
+    Optional<TimetableFrame> findByUserIdAndSemesterAndIsMain(Integer userId, String semester, boolean isMain);
 
     default TimetableFrame getById(Integer id) {
         return findById(id)
             .orElseThrow(() -> TimeTableNotFoundException.withDetail("timetable_frame_id : " + id));
     }
 
-    default TimetableFrame getByUserIdAndSemester(Integer userId, String semester) {
-        return findByUserIdAndSemester(userId, semester)
+    default TimetableFrame getByUserIdAndSemester(Integer userId, String semester, boolean isMain) {
+        return findByUserIdAndSemesterAndIsMain(userId, semester, isMain)
             .orElseThrow(() -> TimeTableNotFoundException.withDetail("userId: " + userId));
     }
 }
