+++ src/main/java/in/koreatech/koin/domain/timetable/repository/TimetableLectureRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.domain.timetable.model.TimetableLecture;
 
 public interface TimetableLectureRepository extends Repository<TimetableLecture, Integer> {
+
     TimetableLecture save(TimetableLecture timetableLecture);
 
     List<TimetableLecture> findAllByTimetableFrameId(Integer timetableFrameId);
